+++ CHANGELOG.md
@@ -13,5 +13,7 @@
 - Fixed missing `package_data` in `setup.py`.
 - Added method to `EnergyCalculator` for saving metadata into a CSV file.
 
-## [0.4.1] - 2021-11-14
+## [0.5.0] - 2021-11-14
 - Renamed `Forcefield.omm2ff` into `Forcefield.map_mda2ff` and made it assert that the mapped atom is known
+- PDB adjustments are now done at the `EnergyCalculator.with_prot_ff` and `EnergyCalculator.with_rna_ff` constructors instead of doing it when calculating the energies.
+- A single `EnergyCalculator.calc_energies` method is now used instead.